+++ src/App.js
@@ -1,12 +1,12 @@
 import React, { Suspense, lazy } from 'react'
 import styled from 'styled-components/macro'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
-import Home from './components/pages/Home/Home'
-import Footer from './components/Footer/Footer'
-import NotFound from './components/pages/NotFound/NotFound'
+import Home from './components/pages/Home'
+import Footer from './components/Footer'
+import NotFound from './components/pages/NotFound'
 
-const Details = lazy(() => import('./components/pages/Details/Details'))
-const About = lazy(() => import('./components/pages/About/About'))
+const Details = lazy(() => import('./components/pages/Details'))
+const About = lazy(() => import('./components/pages/About'))
 
 export default function App() {
   return (
