+++ src/components/pages/Details.js
